Better UI style, new widgets, started settings implementation. Log-in/out works.

+++ test/widget_test.dart
@@ -48,7 +48,7 @@ void main() {
   // });
 
   testWidgets('Scaffold with backdrop drawer responds to menu taps', (WidgetTester tester) async {
-    int drawerWidth = 500;
+    double drawerWidth = 500;
 
     // Build scaffold
     await tester.pumpWidget(
